+++ lib/pgmq/client/topics.rb
@@ -114,8 +114,9 @@ module PGMQ
       # @param routing_key [String] dot-separated routing key
       # @param messages [Array<String>] array of message payloads as JSON strings
       # @param headers [Array<String>, nil] optional array of headers as JSON strings
-      # @param delay [Integer, Time] delay in seconds before messages become visible, or an absolute
-      #   Time at which the messages become visible (PGMQ v1.10.0+)
+      # @param delay [Numeric, Time] delay in seconds before messages become visible (integer or float),
+      #   or an absolute Time at which the messages become visible, including ActiveSupport::TimeWithZone
+      #   (PGMQ v1.10.0+)
       # @return [Array<Hash>] array of hashes with :queue_name and :msg_id
       #
       # @example Batch topic send
@@ -136,7 +137,7 @@ module PGMQ
           encoder = PG::TextEncoder::Array.new
           encoded_messages = encoder.encode(messages)
 
-          if headers && !delay.is_a?(Integer)
+          if headers && !delay.is_a?(Numeric)
             encoded_headers = encoder.encode(headers)
             conn.exec_params(
               "SELECT * FROM pgmq.send_batch_topic($1::text, $2::jsonb[], $3::jsonb[], $4::timestamptz)",
@@ -148,7 +149,7 @@ module PGMQ
               "SELECT * FROM pgmq.send_batch_topic($1::text, $2::jsonb[], $3::jsonb[], $4::integer)",
               [routing_key, encoded_messages, encoded_headers, delay]
             )
-          elsif !delay.is_a?(Integer)
+          elsif !delay.is_a?(Numeric)
             conn.exec_params(
               "SELECT * FROM pgmq.send_batch_topic($1::text, $2::jsonb[], $3::timestamptz)",
               [routing_key, encoded_messages, delay.to_time.utc.iso8601(6)]
